+++ 40-yale-asyncvlsi-act.sh
@@ -3,14 +3,13 @@
 echo "#############################"
 echo "# act"
 
-if [ ! -d $VLSI_TOOLS_SRC ]; then
-cd $EDA_SRCDIR
-  git clone https://github.com/asyncvlsi/act.git
-  mv act $VLSI_TOOLS_SRC
-fi
 cd $VLSI_TOOLS_SRC
-git reset --hard && git checkout master && git pull
 
+if [ -z $FORCE_GIT_RESET ]; then
+   git reset --hard && git checkout master && git pull
+else
+   git pull
+fi
 ./configure $ACT_HOME CFLAGS="-I${ACT_HOME}/include -L${ACT_HOME}/lib" || exit 1
 make realclean 2> /dev/null
 ./build || exit 1
